mechanic_ai/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anaconda-anon-usage @ file:///C:/b/abs_76_7r7k0xr/croot/anaconda-anon-usage_1743190159829/work
annotated-types @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/annotated-types_1729043709318/work
anyio==4.12.0
appdirs==1.4.4
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.10.0
async-generator==1.10
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
base58==1.0.3
bitarray==3.8.0
blinker==1.9.0
boltons @ file:///C:/b/abs_45_52ughkz/croot/boltons_1737061711836/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools==6.2.2
certifi @ file:///C:/b/abs_2bz30lzohz/croot/certifi_1754570654565/work/certifi
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
ckzg==2.1.5
click==8.3.1
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1729036581634/work
conda @ file:///C:/b/abs_95xxia215w/croot/conda_1743715315076/work/conda-src
conda-anaconda-telemetry @ file:///C:/b/abs_8fisg12wdf/croot/conda-anaconda-telemetry_1744662558247/work
conda-anaconda-tos @ file:///C:/b/abs_74fnpuprm8/croot/conda-anaconda-tos_1744823887798/work
conda-content-trust @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/conda-content-trust_1729088072778/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_bdz9vbvbh2/croot/conda-package-streaming_1731366449946/work
constantly==23.10.4
cryptography==46.0.3
cytoolz==1.1.0
daphne==4.2.1
distro @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/distro_1729059153117/work
Django @ file:///C:/b/abs_732pccxzov/croot/django_1745263601709/work
django-cors-headers==4.9.0
django-filter==25.1
django-tailwind==4.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
Faker==37.5.3
Flask==3.1.2
frozendict @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/frozendict_1729090952168/work
frozenlist==1.8.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/idna_1729038997316/work
importlib_metadata==8.7.0
incremental==24.7.2
ipfs_node==0.1.11
ipfs_tk==0.1.9
ipfs_toolkit==0.5.33
ipfshttpclient==0.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jsonpatch @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/jsonpatch_1729054776004/work
jsonpointer==2.1
libmambapy @ file:///C:/b/abs_627vsv8bhu/croot/mamba-split_1734469608328/work/libmambapy
Markdown==3.9
markdown-it-py @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/markdown-it-py_1729039967492/work
MarkupSafe==3.0.3
mdurl @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mdurl_1729039154828/work
menuinst @ file:///C:/b/abs_fblttj5gp1/croot/menuinst_1738943438301/work
morphys==1.0
msgpack==1.1.2
multiaddr==0.1.0
multidict==6.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
netaddr==1.3.0
openai==2.20.0
outcome==1.3.0.post0
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
parsimonious==0.10.0
platformdirs @ file:///C:/b/abs_ddh15014or/croot/platformdirs_1744273060660/work
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
propcache==0.4.1
psutil==7.1.3
psycopg==3.2.9
psycopg2 @ file:///C:/b/abs_a4bewkfd4i/croot/psycopg2_1744919806261/work
py-cid==0.3.1
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==2.0.1
py-solc-x==2.0.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.23.0
pydantic @ file:///C:/b/abs_27dx58x550/croot/pydantic_1734736090499/work
pydantic_core @ file:///C:/b/abs_bdosz7qwys/croot/pydantic-core_1734726071532/work
pygame==2.6.1
Pygments @ file:///C:/b/abs_e4bg5vh5j_/croot/pygments_1744667628203/work
PyJWT==2.10.1
pymultihash==0.8.2
PyMySQL==1.1.1
pyOpenSSL==25.3.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1729039320841/work
python-baseconv==1.2.2
python-decouple==3.8
python-dotenv==1.1.1
pyunormalize==17.0.0
pywin32==311
redis==7.0.0
regex==2025.11.3
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
rich @ file:///C:/b/abs_8b92fzmygg/croot/rich_1732638991774/work
rlp==4.1.0
ruamel.yaml @ file:///C:/b/abs_0dn1nhdq75/croot/ruamel.yaml_1745960327187/work
ruamel.yaml.clib @ file:///C:/b/abs_475flulhy_/croot/ruamel.yaml.clib_1745937166245/work
service-identity==24.2.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse @ file:///C:/Users/dev-admin/buildout/croot/sqlparse_1739565583021/work
termcolor==3.2.0
toolz==1.1.0
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
trio==0.32.0
trio-typing==0.10.0
truststore @ file:///C:/b/abs_494cm143zh/croot/truststore_1736550137835/work
Twisted==25.5.0
txaio==25.9.2
types-requests==2.32.4.20250913
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
varint==1.0.2
web3==7.14.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
win_inet_pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1729038644686/work
yarl==1.22.0
zipp==3.23.0
zope.interface==8.0.1
zstandard @ file:///C:/b/abs_31t8xmrv_h/croot/zstandard_1731356578015/work
